# Use iterators for converting RGB to BGRA in the Rust PNG benchmark