@landaire 2021-04-07
I force pushed a new change which correctly does the RGB to BGRA encoding. po8's patch just copied from src to dst without taking into account that `R` and `B` are in opposite positions. The benchmarks were updated accordingly.